AnalyticsView is now successfully displayed
