+++ infrastructure/site/lib/screens/analytics0/view/analytics_view.dart
@@ -1,0 +1,55 @@
+import 'package:admin/api_repository.dart';
+import 'package:admin/screens/analytics0/analytics.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+
+class AnalyticsRoute extends StatelessWidget {
+  const AnalyticsRoute({Key? key, required this.repository}) : super(key: key);
+
+  final ApiRepository repository;
+
+  @override
+  Widget build(BuildContext context) {
+    return RepositoryProvider.value(
+      value: repository,
+      child: BlocProvider(
+        create: (_) => AnalyticsBloc(repository: repository)
+          ..add(const AnalyticsStarted(operation: 'count')),
+        child: const AnalyticsView(),
+      ),
+    );
+  }
+}
+
+class AnalyticsView extends StatelessWidget {
+  const AnalyticsView({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return BlocBuilder<AnalyticsBloc, AnalyticsState>(
+      builder: (context, state) => _handleLoadEvents(context, state),
+    );
+  }
+
+  Widget _handleLoadEvents(BuildContext context, AnalyticsState state) {
+    return Padding(
+      padding: EdgeInsets.all(16),
+      child: Center(child: analyticsFromState(state)),
+    );
+  }
+
+  Widget analyticsFromState(AnalyticsState state) {
+    if (state is AnalyticsInitial) {
+      return Text('No analytics yet.');
+    } else if (state is AnalyticsInProgress) {
+      return CircularProgressIndicator();
+    } else if (state is AnalyticsFailure) {
+      return Text('Error fetching analytics.');
+    } else if (state is AnalyticsSuccess) {
+      return Text('${state.result}');
+    } else {
+      assert(false);
+      return ErrorWidget('Error State');
+    }
+  }
+}
